Refactor Additional Mountains on Ascent Modal for upcoming improvements
